==> src/2-PlotResults.Rmd <==
---
title: "Explore results of the ocean climate novelty"
author: "Katie Lotterhos"
date: "6/27/2019"
output: html_document
---
  
setwd("~/Documents/Github/OceanClimateNovelty/src")  
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
source("0-Novelty_Oceans_Functions.R")
```



```{r load data}

dat_Nov <- fread("../results_annual/SigmaD.csv")
dat1 <- fread("/Users/lotterhos/Google Drive/katie_research/2018-OceanClimateNovelty/Data/Lotterhos/Katie_Temp_Arag_1800_2000.txt", sep = ",")
dat4.5 <- fread("/Users/lotterhos/Google Drive/katie_research/2018-OceanClimateNovelty/Data/Lotterhos/Katie_Temp_Arag_2070_2100_RCP45.txt", sep = ",")
dat8.5 <- fread("/Users/lotterhos/Google Drive/katie_research/2018-OceanClimateNovelty/Data/Lotterhos/Katie_Temp_Arag_2070_2100_RCP85.txt", sep = ",")


```

# Plot distribution of temperature and omega_arag at 1800, 2000, and 2100
```{r, eval=FALSE, echo=FALSE}

# This code generates a plot with a hard-to-see pattern. 

# plot(dat8.5$SST[dat1$Year==1800], dat8.5$Arag[dat1$Year==1800], pch=19, col=adjustcolor("black", 0.5))
# 
# points(dat8.5$SST[dat1$Year==2000], dat8.5$Arag[dat1$Year==2000], pch=17, col=adjustcolor("lightblue", 0.2))
# 
# points(dat1$SST[dat8.5$Year==2100], dat8.5$Arag[dat1$Year==2100], pch=22, col=adjustcolor("darkorange", 0.1), bg=adjustcolor("yellow", 0.2))
```

# Make Temp vs. Omega Plot
Subset to points with unique x,y values:
```{r}
summary(dat1$SST)
summary(dat1$Arag)

ab=matrix(c(-5, 40, 0,12), nrow=2, byrow=TRUE)
nbin = c(100,100)
x <- seq(ab[1,1], ab[1,2], length.out = nbin[1])
y <- seq(ab[2,1], ab[2,2], length.out = nbin[2])

thresh <- 1

setup_plot <- function(x){
  x_yo <- x
  x_yo[x>1000] <- 4
  x_yo[x<=1000 & x>100 ] <- 3
  x_yo[x<=100 & x>1] <- 2
  return(x_yo)
}
  
```

```{r grouping based on climate normals, echo=FALSE, eval=FALSE}
# ### 1800 layer
# dat1norm_1800 <- dat1 %>% filter(Year<1850) %>% group_by(No) %>%
#   summarise_each(mean)
# head(dat1norm)
# dim(dat1norm)
#   y1800 <- bin2(cbind(dat1norm_1800$SST, dat1norm_1800$Arag) , ab, nbin)
#   z_1800 <- y1800$nc
#   z_1800[z_1800<=thresh] <- 0 # convert to presence/absence
#   z_1800[z_1800>thresh] <- 1 # convert to presence/absence
#   
# ## 2000 layer  
# dat1norm_2000 <- dat1 %>% filter(Year>1950) %>% group_by(No) %>%
#   summarise_each(mean)  
#   y2000 <- bin2(cbind(dat1norm_2000$SST, dat1norm_2000$Arag) , ab, nbin)
#   z_2000 <- y2000$nc
#   z_2000[z_2000<=thresh] <- 0 # convert to presence/absence
#   z_2000[z_2000>thresh] <- 1 # convert to presence/absence
#  
# ## 2100 layer 8.5
#   dat1norm_8.5 <- dat8.5  %>% group_by(No) %>%
#   summarise_each(mean)  
#   y2100_8.5 <- bin2(cbind(dat1norm_8.5$SST, dat1norm_8.5$Arag) , ab, nbin)
#   z_2100_8.5 <- y2100_8.5$nc
#   z_2100_8.5[z_2100_8.5<=thresh] <- 0 # convert to presence/absence
#   z_2100_8.5[z_2100_8.5>thresh] <- 1 # convert to presence/absence 
#   
# ## 2100 layer 4.5
#     dat1norm_4.5 <- dat4.5  %>% group_by(No) %>%
#   summarise_each(mean) 
#   y2100_4.5 <- bin2(cbind(dat1norm_4.5$SST, dat1norm_4.5$Arag) , ab, nbin)
#   z_2100_4.5 <- y2100_4.5$nc
#   z_2100_4.5[z_2100_4.5<=thresh] <- 0 
#   z_2100_4.5[z_2100_4.5>thresh] <- 1 # convert to presence/absence     
```

# Temp vs. omega binning

```{r grouping based on all climates}
## 1800 layer
  y1800 <- bin2(cbind(dat1$SST[dat1$Year<1850], dat1$Arag[dat1$Year<1850]) , ab, nbin)
  z_1800 <- y1800$nc
  #z_1800[z_1800<=thresh] <- 0 # convert to presence/absence
  #z_1800[z_1800>thresh] <- 1 # convert to presence/absence

## 2000 layer
  y2000 <- bin2(cbind(dat1$SST[dat1$Year>1950], dat1$Arag[dat1$Year>1950]) , ab, nbin)
  z_2000 <- y2000$nc
  #z_2000[z_2000<=thresh] <- 0 # convert to presence/absence
  #z_2000[z_2000>thresh] <- 1 # convert to presence/absence
 
## 2100 layer 8.5
  y2100_8.5 <- bin2(cbind(dat8.5$SST, dat8.5$Arag) , ab, nbin)
  z_2100_8.5 <- y2100_8.5$nc
  #z_2100_8.5[z_2100_8.5<=thresh] <- 0 # convert to presence/absence
  #z_2100_8.5[z_2100_8.5>thresh] <- 1 # convert to presence/absence   
  
## 2100 layer 4.5
  y2100_4.5 <- bin2(cbind(dat4.5$SST, dat4.5$Arag) , ab, nbin)
  z_2100_4.5 <- y2100_4.5$nc
  #z_2100_4.5[z_2100_4.5<=thresh] <- 0 
  #z_2100_4.5[z_2100_4.5>thresh] <- 1 # convert to presence/absence   
```

# Temp vs. pH binning
Subset to points with unique x,y values:
```{r}
summary(c(dat1$SST, dat8.5$SST))
summary(c(dat1$pH, dat8.5$pH))

ab2=matrix(c(-5, 40, 7,9), nrow=2, byrow=TRUE)
nbin = c(100,100)
xb <- seq(ab2[1,1], ab2[1,2], length.out = nbin[1])
yb <- seq(ab2[2,1], ab2[2,2], length.out = nbin[2])

## 1800 layer
  y1800 <- bin2(cbind(dat1$SST[dat1$Year<1850], dat1$pH[dat1$Year<1850]) , ab2, nbin)
  z_1800b <- y1800$nc
  #z_1800b[z_1800b>0] <- 1 # convert to presence/absence

## 2000 layer
  y2000 <- bin2(cbind(dat1$SST[dat1$Year>1950], dat1$pH[dat1$Year>1950]) , ab2, nbin)
  z_2000b <- y2000$nc
  #z_2000b[z_2000b>0] <- 1 # convert to presence/absence
 
## 2100 layer 8.5
  y2100_8.5 <- bin2(cbind(dat8.5$SST, dat8.5$pH) , ab2, nbin)
  z_2100_8.5b <- y2100_8.5$nc
  #z_2100_8.5b[z_2100_8.5b>0] <- 1 # convert to presence/absence   
  
## 2100 layer 4.5
  y2100_4.5 <- bin2(cbind(dat4.5$SST, dat4.5$pH) , ab2, nbin)
  z_2100_4.5b <- y2100_4.5$nc
  #z_2100_4.5b[z_2100_4.5b>0] <- 1 # convert to presence/absence 
  
```  

### Make Temp vs. Omega and Temp vs. pH plot
```{r}
### Make Temp vs. Omega plot ####  
pdf("../figures/TempVsOmega2vspH.pdf", width=8, height=7)  
par(mar=c(2,2,0.5,0.5), mfrow=c(2,2), oma=c(2,2,0.5,0.5))  
# 1800 to 2000  

  get_cols <- function(string){
    c(rgb(0,0,0,0), 
      adjustcolor(string,0.2),
      adjustcolor(string,0.4),
      adjustcolor(string,0.6),
      adjustcolor(string,0.8)
      )
  }
  get_cols("dodgerblue4")
  
  image(x, y, setup_plot(z_1800), col=get_cols("blue"),
        xlab="Temperature (C)",
        ylab="Aragonite Saturatation State",
        bty="l",
        xlim=c(-5, 60),
        las=1)
  image(x, y, setup_plot(z_2000), add=TRUE, col=get_cols("goldenrod"))
  contour(x,y, setup_plot(z_1800), nlevels=2, add=TRUE, col=get_cols("blue")[c(4,2)], drawlabels=FALSE)
  contour(x,y, setup_plot(z_2000), nlevels=2, add=TRUE, col=get_cols("goldenrod")[c(5,3)], drawlabels=FALSE)
  
  text(0, 11, "A", cex=2)
  text(20, 7.5, "1800 disappearing in 2000", col="dodgerblue4")  
  lines(c(20,25), c(7, 4.6), lwd=3, col="dodgerblue4")
  lines(c(20,5.3), c(7, 4.1), lwd=3, col="dodgerblue4")
   abline(1,0, lty=2, col=rgb(0,0,0,0.7))
  text(30, 2.0, "Novel in 2000", col="goldenrod4", adj=0)  
  lines(c(29,28), c(2.0, 3.0), lwd=3, col="goldenrod4")
  lines(c(29,17), c(2.0, 1.9), lwd=3, col="goldenrod4")
  legend(40, 12, c("1800", "2000"), fill=c(adjustcolor("blue", 0.8), 
      adjustcolor("goldenrod", 0.8)), bty="n", adj=0, cex=1.5)
  mtext("Aragonite Saturation State", side=2, outer=TRUE, adj=0.95, cex=1.4)
    
# 2000 to 2100_8.5  
  image(x, y, setup_plot(z_2000), col=get_cols("goldenrod"),
      xlab="Temperature (C)",
        ylab="",
        bty="l",
        xlim=c(-5, 60),
        las=1
        )
  image(x, y, setup_plot(z_2100_8.5), add=TRUE, col=get_cols("darkmagenta"))
  contour(x,y, setup_plot(z_2100_8.5), nlevels=2, add=TRUE, col=get_cols("darkmagenta")[c(4,2)], drawlabels=FALSE)
  contour(x,y, setup_plot(z_2000), nlevels=2, add=TRUE, col=get_cols("goldenrod")[c(5,3)], drawlabels=FALSE)
  text(20, 6.5, "2000 disappearing in\n2100 RCP 8.5", col="goldenrod4")  
  text(0, 11, "B", cex=2)
  lines(c(20,20), c(5.5, 3.2), lwd=3, col="goldenrod4")
  lines(c(20,10), c(5.5, 2.2), lwd=3, col="goldenrod4")
  abline(1,0, lty=2, col=rgb(0,0,0,0.7))
  text(36, 2, "Novel in 2100\nRCP 8.5", col="darkmagenta", adj=0)  
  lines(c(35,32), c(2, 2.5), lwd=3, col="darkmagenta")
  lines(c(35,24), c(2, 1.9), lwd=3, col="darkmagenta")
  legend(40, 12, c("2000", "2100"), fill=c(adjustcolor("goldenrod", 0.8), 
         adjustcolor("darkmagenta", 0.8)), bty="n", adj=0, cex=1.5)

### PH ######
# 1800 to 2000  PH
  image(xb, yb, setup_plot(z_1800b), col=get_cols("blue"),
        xlab="Temperature (C)",
        ylab="pH",
        bty="l",
        xlim=c(-5, 60),
        las=1
        )
  image(xb, yb,setup_plot(z_2000b), add=TRUE, col=get_cols("goldenrod"))
  contour(xb,yb, setup_plot(z_1800b), nlevels=2, add=TRUE, col=get_cols("blue")[c(4,2)], drawlabels=FALSE)
  contour(xb,yb, setup_plot(z_2000b), nlevels=2, add=TRUE, col=get_cols("goldenrod")[c(5,3)], drawlabels=FALSE)
  text(0, 8.9, "C", cex=2)
  #text(20, 7.5, "1800 disappearing in 2000", col="dodgerblue4")  
  #lines(c(20,25), c(7, 4.6), lwd=3, col="dodgerblue4")
  #lines(c(20,5.3), c(7, 4.1), lwd=3, col="dodgerblue4")
  # abline(1,0, lty=2, col=rgb(0,0,0,0.7))
  #text(30, 2.0, "Novel in 2000", col="goldenrod4", adj=0)  
  #lines(c(29,28), c(2.0, 3.0), lwd=3, col="goldenrod4")
  #lines(c(29,17), c(2.0, 1.9), lwd=3, col="goldenrod4")
 # legend(40,8.9, c("1800", "2000"), fill=c(adjustcolor("dodgerblue4", 0.8), 
   #   adjustcolor("goldenrod", 0.8)), bty="n", adj=0)
  mtext("pH", side=2, outer=TRUE, adj=0.25,  cex=1.5, line=0)
    
# 2000 to 2100_8.5  PH
  image(xb, yb, setup_plot(z_2000b), col=get_cols("goldenrod"),
        xlab="Temperature (C)",
        ylab="",
        bty="l",
        xlim=c(-5, 60),
        las=1
        )
  image(xb, yb, setup_plot(z_2100_8.5b), add=TRUE, col=get_cols("darkmagenta"))
  contour(xb,yb, setup_plot(z_2100_8.5b), nlevels=2, add=TRUE, col=get_cols("darkmagenta")[c(4,2)], drawlabels=FALSE)
  contour(xb,yb, setup_plot(z_2000b), nlevels=2, add=TRUE, col=get_cols("goldenrod")[c(5,3)], drawlabels=FALSE)
  #text(20, 6.5, "2000 disappearing in\n2100 RCP 8.5", col="goldenrod4")  
  text(0, 8.9, "D", cex=2)
  #lines(c(20,20), c(5.5, 3), lwd=3, col="goldenrod4")
  #lines(c(20,10), c(5.5, 2), lwd=3, col="goldenrod4")
  #abline(1,0, lty=2, col=rgb(0,0,0,0.7))
  #text(36, 2, "Novel in 2100\nRCP 8.5", col="darkmagenta", adj=0)  
  #lines(c(35,32), c(2, 2.5), lwd=3, col="darkmagenta")
  #lines(c(35,24), c(2, 1.9), lwd=3, col="darkmagenta")
 # legend(40, 8.9, c("2000", "2100"), fill=c(adjustcolor("goldenrod", 0.8), 
  #       adjustcolor("darkmagenta", 0.5)), bty="n", adj=0)
  mtext("Temperature", outer=TRUE, side=1, line=1, cex=1.5)
dev.off()  
```


## Visualize pH, temp, and Arag
```{r}
dat_2000 <- dat1 %>% filter(Year>1950)
subsamp <- sample(1:nrow(dat_2000), 30000, replace=FALSE)


p1 <- ggplot(dat_2000[subsamp,], aes(pH, Arag)) + 
  geom_point(aes(color=SST), alpha=0.3, pch=20,  size=3) + 
  scale_fill_gradient(low = "blue",high = "red")  +
  scale_color_continuous(low = "blue",high = "red") +
  xlab("pH") + ylab("Aragonite Saturation State") + 
  theme(legend.position = "right") +
  theme_classic()

p2 <- ggplot(dat_2000[subsamp,], aes(pH, Arag)) + 
  geom_point(aes(color=Lat), alpha=0.3, pch=20,  size=3) + 
  #scale_fill_gradient2(low = "green",high = "blue", mid="grey")
  scale_fill_gradient(low = "blue",high = "black")  +
  scale_color_continuous(low="yellow", high="purple") +
  xlab("pH") + ylab("Aragonite Saturation State") + 
  theme(legend.position = "right") +
  theme_classic()


png("../figures/pHvsOmega.png", width=4, height=7, units="in", res=600)
grid.arrange(p1, p2, nrow=2)
dev.off()

dat_2000 %>% filter(pH<7.8) %>% select(Lat) %>% summary
```


## Subset to specific places on map:

Questions: 
For the future Antartic, where is the climate going to come from today?


```{r}
tmp <- data.frame(x=dat1$Lon, y=dat1$Lat, z=dat1$SST)
tmp2 <- tmp %>% group_by(x, y) %>% summarise(z=mean(z))
forPlot <- acast(tmp2, x~y, value.var="z")
dim(forPlot)
head(forPlot)

image.plot(y=as.numeric(colnames(forPlot)),x=as.numeric(rownames(forPlot)), forPlot)
polygon(x=c(50,50,200,200), y=c(-25,25,25,-25)) # IndoPacific
polygon(x=c(20, 380, 380, 20), y=c(-90, -90, -66.5, -66.5))
polygon(x=c(20, 380, 380, 20), y=c(90, 90, 66.5, 66.5))

summary(tmp2$x)

### Subset
head(dat_Nov)
x <- dat_Nov
cond_indop <- function(x){x$Lon<200 & x$Lon>50 & 
  abs(x$Lat) < 25}
cond_antar <- function(x){x$Lat < -66.5}
cond_artic <- function(x){x$Lat > 66.5}


ab=matrix(c(-5, 40, 0,12), nrow=2, byrow=TRUE)
nbin = c(100,100)
x <- seq(ab[1,1], ab[1,2], length.out = nbin[1])
y <- seq(ab[2,1], ab[2,2], length.out = nbin[2])
## GetEnvel function ####
PlotEnvel <- function(datf, col2){
  a1 <- datf %>% select(SST, Arag) 
  a1 <- bin2(cbind(a1$SST, a1$Arag), ab, nbin)
  a2 <- a1$nc
  a2[a2>0] <- 1
  image(x,y, a2, add=TRUE, col=c(rgb(0,0,0,0), col2))  
}
```

```{r}
## Plot climate envelopes for specific regions
plot(c(-5, 40), c(0, 12), type="n")
IndoPac_2000 <- dat1 %>% filter(Lon<200 & Lon>50 & abs(Lat) < 25 & Year > 1950 & Year < 2010)
PlotEnvel(IndoPac_2000, "darkgreen")
IndoPac_2100_8.5 <- dat8.5 %>% filter(Lon<200 & Lon>50 & 
  abs(Lat) < 25)
PlotEnvel(IndoPac_2100_8.5, adjustcolor("darkgreen", 0.5))

```

```{r}

plot(c(-5, 15), c(0, 4), type="n")
Art_2000 <- dat1 %>% filter(Lat > 66.5 & Year > 1950 & Year < 2010)
PlotEnvel(Art_2000, adjustcolor("red", 0.6))
Art_2100_8.5 <- dat8.5 %>% filter(Lat > 66.5)
PlotEnvel(Art_2100_8.5, adjustcolor("red", 0.3))

Ant_2000 <- dat1 %>% filter(Lat < -66.5 & Year > 1950 & Year < 2010)
PlotEnvel(Ant_2000, adjustcolor("blue", 0.7))
Ant_2100_8.5 <- dat8.5 %>% filter(Lat < -66.5)
PlotEnvel(Ant_2100_8.5, adjustcolor("yellow", 0.5))

```
